Revenue changed from Int to Double; Tests implemented (rebased) (#53)
* Revenue changed from Int to Double; Tests implemented
* Updated tests

+++ AutocompleteClient/FW/Logic/Request/CIOTrackConversionData.swift
@@ -16,13 +16,13 @@ public struct CIOTrackConversionData: CIORequestData {
     public let searchTerm: String
     public let itemID: String
     public var sectionName: String?
-    public let revenue: Int?
+    public let revenue: Double?
     
     public var url: String{
         return String(format: Constants.TrackConversion.format, Constants.Track.baseURLString, self.searchTerm.addingPercentEncoding(withAllowedCharacters: .urlPathAllowed)!)
     }
 
-    public init(searchTerm: String, itemID: String, sectionName: String? = nil, revenue: Int? = nil) {
+    public init(searchTerm: String, itemID: String, sectionName: String? = nil, revenue: Double? = nil) {
         self.searchTerm = searchTerm
         self.itemID = itemID
         self.sectionName = sectionName
